requirements.txt
streamlit==1.28.0
google-generativeai==0.3.0
folium==0.14.0
streamlit-folium==0.15.0
python-dotenv==1.0.0
reportlab==4.0.4
requests==2.31.0
plotly==5.17.0
pandas==2.1.0
numpy==1.24.3